Chore:데이터 패칭 커스텀 훅 제거
